[반영내용]Record Explorer 1. 파일 업로드 중 Abort 클릭시 DropZone 표시 삭제 처리

+++ base/src/main/java/rmsoft/ams/seoul/rc/rc005/vo/Rc00502VO.java
@@ -36,7 +36,7 @@ public class Rc00502VO extends BaseVO {
 
     private String sourceSystemUuid;
 
-    private int contentsSize;
+    private long contentsSize;
 
     private String fileFormatUuid;
